+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Ember Flexberry Designer Changelog
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+### Added
+* Added SignalR service for real-time notifications about diagram updates
+* Added diagram updated notification popup
+
 ## [1.1.2-beta.9] - 2026-03-24
 ### Fixed
 * Fix share